Ok I did a lot of reading today. I'm going to split each FET into a sheet, so I can add the external circuitry more easily.
I may need to source a different FET. The one I selected has a good EAS, but the rating is only for 11A. and Ideally I would be able to do 19 or 20A. I'm sure I will never do that much current , but relying on the mosfet to handle things seems better than adding an additional tvs diode. Snubber is the way to go for sure to keep things cool.